Fixed Double.compareTo(Object) problem by casting the Object to a Double
git-svn-id: https://jep.svn.sourceforge.net/svnroot/jep/trunk/jep@511 216fd33d-161a-0410-b2e1-d58788b044ed

+++ src/org/lsmp/djep/sjep/PConstant.java
@@ -120,12 +120,12 @@ public class PConstant extends AbstractPNode {
 	}
 	public boolean isPositive() {
 		try	{
-			return ((Double) value).compareTo(pc.zero) > 0;
+			return ((Double) value).compareTo((Double) pc.zero) > 0;
 		} catch(Exception e) { return false; }
 	}
 	public boolean isNegative() {
 		try	{
-			return ((Double) value).compareTo(pc.zero) < 0;
+			return ((Double) value).compareTo((Double) pc.zero) < 0;
 		} catch(Exception e) { return false; }
 	}
 	public boolean isInteger() {
